doc: stylish stash.md, #24717
