+++ akka-docs/src/main/paradox/typed/stash.md
@@ -30,7 +30,7 @@ Scala
 :  @@snip [StashDocSpec.scala](/akka-actor-typed-tests/src/test/scala/docs/akka/typed/StashDocSpec.scala) { #stashing }
 
 Java
-:  @@snip [StashDocTest.java](/akka-actor-typed-tests/src/test/java/jdocs/akka/typed/StashDocTest.java) {
+:  @@snip [StashDocTest.java](/akka-actor-typed-tests/src/test/java/jdocs/akka/typed/StashDocSample.java) {
   #import
   #db
   #stashing
